device-manager: move the generic plugin tests
Move the lifecycle device plugin for the generic device plugin in the
device plugin base.
